implement list view in events and start managing post requests
Got fairly familiar with context data which is forwarded to html. This 
should be no problem in the future. 
Next step is to create the Participation queries needed for processing 
get and post requests. But after this I should have a pretty nice 
teamplanbuch functionality. THen I could worry about the details